Add session management and JWT validation enhancements
- Implemented a new endpoint in SessionApi to check active sessions.
- Enhanced JwtFilter to validate JWT tokens and set the username attribute.
- Updated JWTHelper to include issuer and subject validation for JWT tokens.
- Added session checks in BlogService methods to ensure actions are performed only with an active session.
- Introduced isSessionActive and getUsername methods in SessionService for session management.

+++ src/main/java/net/ausiasmarch/persutil/service/BlogService.java
@@ -20,6 +20,9 @@ public class BlogService {
     @Autowired
     AleatorioService oAleatorioService;
 
+    @Autowired
+    SessionService oSessionService;
+
     ArrayList<String> alFrases = new ArrayList<>();
 
     public BlogService() {
@@ -42,6 +45,11 @@ public class BlogService {
     }
 
     public Long rellenaBlog(Long numPosts) {
+
+        if (!oSessionService.isSessionActive()) {
+            throw new RuntimeException("No active session");
+        }
+
         for (long j = 0; j < numPosts; j++) {
             // crea entity blog y la rellana con datos aleatorios
             BlogEntity oBlogEntity = new BlogEntity();
@@ -102,6 +110,9 @@ public class BlogService {
     }
 
     public Long create(BlogEntity blogEntity) {
+        if (!oSessionService.isSessionActive()) {
+            throw new RuntimeException("No active session");
+        }
         blogEntity.setFechaCreacion(LocalDateTime.now());
         blogEntity.setFechaModificacion(null);
         oBlogRepository.save(blogEntity);
@@ -109,6 +120,9 @@ public class BlogService {
     }
 
     public Long update(BlogEntity blogEntity) {
+        if (!oSessionService.isSessionActive()) {
+            throw new RuntimeException("No active session");
+        }
         BlogEntity existingBlog = oBlogRepository.findById(blogEntity.getId())
                 .orElseThrow(() -> new RuntimeException("Blog not found"));
         existingBlog.setTitulo(blogEntity.getTitulo());
@@ -120,6 +134,9 @@ public class BlogService {
     }
 
     public Long delete(Long id) {
+        if (!oSessionService.isSessionActive()) {
+            throw new RuntimeException("No active session");
+        }
         oBlogRepository.deleteById(id);
         return id;
     }
